refactor(agglayer): move ProofData and LeafData to types module
ProofData, LeafData, and the associated type aliases (SmtNode, ExitRoot,
LeafValue, CgiChainHash) are not specific to CLAIM-note construction.
Move them to a new types module so they can be used across the agglayer
crate without implying a dependency on claim_note.

Closes #2907
